Fixes test for CI
the test was trying to check the behaviour when the target dir caould not be
created. It used a readonly parent dir to do so.

CI however runs in a docker container as root, and root can create subdirs in
readonly directories.

This commit changes the strategy by creating a file with the same name as the
directory to be created, so that this fails.