bug: added an exception for the legacy R66 certificate
The legacy R66 certificate, which is packaged with Waarp-R66, is rejected by the gateway when used because it is not a valid certificate. So an exception was added in the gateway to allow the usage of this certificate for those who need it. This exception is turned off by default, and can be activated using an environment variable.